re_intuit_start(): use different var for tmp value
Make the anchored branch more similar to the floating branch by using s to
hold the start position for fbm rather than t.  Should be functionally
equivalent.

Note that on failure in the anchored branch, we leave with t holding a
different value than before, but it shouldn't matter, since the value of t
is only used in the success case.